requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
Deprecated==1.2.14
Django
djangorestframework==3.14.0
firebase-rest-api==1.11.0
gcloud==0.17.0
google-api-core==2.16.1
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
httplib2==0.22.0
idna==3.6
jwcrypto==1.5.1
jws==0.1.3
oauth2client==3.0.0
pkce==1.0.3
proto-plus==1.23.0
protobuf==4.25.2
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome
pyparsing==3.1.1
python-jwt==4.1.0
pytz==2024.1
requests==2.31.0
requests-toolbelt==0.7.0
rsa==4.9
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
stripe==8.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wheel==0.42.0
wrapt==1.16.0
<<<<<<< HEAD
pillow
django-cors-headers
=======
scikit-learn
gradio_client
numpy
openai
firebase-admin
>>>>>>> f312f15a6dc800f12f85fe822c28c3ca622ecaaa
